Remove my own error responder and use the one in Asp.net Core

+++ Common/HttpHelpers/ErrorResponder.cs
@@ -17,7 +17,7 @@ namespace Common.HttpHelpers
         {
             var problem = new ProblemDetails()
             {
-                Status = (long)statusCode,
+                Status = (int)statusCode,
                 Type = type,
                 Title = title,
                 Instance = instance,
